-ensure that we don't rely on stl containters get_allocator, since STL may not support  stateful allocators - fixes issue in using new Pool allocator logic on non-linux  platforms -fixed some compiler warnings
git-svn-id: https://svn.resiprocate.org/rep/resiprocate/branches/b-TKLC-perf_work@9344 ddefafc4-47db-0310-ae44-fa13212b10f2

+++ resip/stack/HeaderFieldValue.hxx
@@ -56,7 +56,7 @@ class HeaderFieldValue
          }
          
          mField=field;
-         mFieldLength=length;
+         mFieldLength=(unsigned int)length;
          mMine=own;
       }
